+++ src/buffer/scroll_view.rs
@@ -105,7 +105,7 @@ pub fn view<'a>(
         max_nick_chars,
         max_prefix_chars,
         ..
-    }) = history.get_messages(&kind.into(), Some(state.limit), &config.buffer)
+    }) = history.get_messages(&kind.into(), Some(state.limit), config)
     else {
         return column![].into();
     };
@@ -392,7 +392,7 @@ impl State {
                             old_messages,
                             new_messages,
                             ..
-                        }) = history.get_messages(&kind.into(), Some(self.limit), &config.buffer)
+                        }) = history.get_messages(&kind.into(), Some(self.limit), config)
                         {
                             if let Some(oldest) = old_messages.iter().chain(&new_messages).next() {
                                 self.limit = Limit::Since(oldest.server_time);
@@ -645,7 +645,7 @@ impl State {
             old_messages,
             new_messages,
             ..
-        }) = history.get_messages(&kind.into(), None, &config.buffer)
+        }) = history.get_messages(&kind.into(), None, config)
         else {
             // We're still loading history, which will trigger
             // scroll_to_backlog after loading. If this is set,
@@ -689,7 +689,7 @@ impl State {
             total,
             old_messages,
             ..
-        }) = history.get_messages(&kind.into(), None, &config.buffer)
+        }) = history.get_messages(&kind.into(), None, config)
         else {
             return Task::none();
         };